+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - \#2772 - URI fields should be links
 - \#2775 - Syntax highlighting for JSON fragments in configuration
 - \#2809 - Health filter in sidebar
+- \#2824 - Add UI option for Docker forcePullImage
 
 ### Changed
 - \#2764 - Make health status bar wider